# Improve file closing performance.